siliconflow-ui/requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
pydub>=0.25.1
python-dotenv>=1.0.0
requests>=2.28.0
pypinyin>=0.48.0
matplotlib>=3.7.0
